+++ tramtracker/types.py
@@ -88,18 +88,30 @@ class TramDeparture(TramTrackerData):
 
     stop_id: int
     """TramTracker code of the stop of this departure"""
-    trip_id: int | None
-    """Trip identifier; currently unused"""
+    trip_id: Literal[0]
+    """Trip identifier; currently unused
+    
+    .. versionchanged:: 0.5.0
+        Changed type hint from ``int | None`` to :class:`Literal[0] <typing.Literal>` as it currently only returns that value
+    """
     route_id: int
     """Route identifier for this departure"""
     route_number: str
     """Public-facing route number for this departure"""
     primary_route_number: str
     """Route number of the main route that this departure belongs to"""
-    vehicle_id: int | None
-    """Identifier of the tram operating this service, as printed on and inside the vehicle; None if information is not currently available"""
-    vehicle_class: Literal["W", "Z3", "A1", "A2", "B2", "C1", "C2", "D1", "D2", "E", "G"] | None
-    """Class/model of the tram operating this service; None if information is not currently available"""
+    vehicle_id: int
+    """Identifier of the tram operating this service, as printed on and inside the vehicle; ``0`` if information is not currently available
+    
+    .. versionchanged:: 0.5.0
+        Value returned as-is from server (instead of substituting ``None`` for ``0``) for consistency with :mod:`ptv_timetable` definitions
+    """
+    vehicle_class: Literal["W", "Z3", "A1", "A2", "B2", "C1", "C2", "D1", "D2", "E", "G", ""]
+    """Class/model of the tram operating this service; empty string if information is not currently available
+    
+    .. versionchanged:: 0.5.0
+        Value returned as-is from server (instead of substituting ``None`` for the empty string) for consistency with :mod:`ptv_timetable` definitions
+    """
     destination: str
     """Destination of this service"""
     tt_available: bool
@@ -116,14 +128,35 @@ class TramDeparture(TramTrackerData):
     """Descriptions of the disruptions affecting this service"""
     has_special_event: bool
     """Whether a special event is affecting or will affect this route"""
-    special_event_message: str | None
-    """Description of the special event"""
+    special_event_message: str
+    """Description of the special event; empty string if none
+    
+    .. versionchanged:: 0.5.0
+        Value returned as-is from server (instead of substituting ``None`` for the empty string) for consistency with :mod:`ptv_timetable` definitions"""
     has_planned_occupation: bool
     """Whether planned service changes are affecting/will affect this route"""
-    planned_occupation_message: str | None
-    """Description of the planned service changes"""
+    planned_occupation_message: str
+    """Description of the planned service changes; empty string if none
+    
+    .. versionchanged:: 0.5.0
+        Value returned as-is from server (instead of substituting ``None`` for the empty string) for consistency with :mod:`ptv_timetable` definitions"""
     estimated_departure: datetime
     """Estimated real-time departure time of this service from this stop"""
+    location: tuple[float, float]
+    """GPS coordinates of the tram's current location; ``(0.0, 0.0)`` if information is not currently available
+
+    .. versionadded:: 0.5.0
+    """
+    occupancy_level: Literal["NO_DATA_AVAILABLE", "EMPTY", "MANY_SEATS_AVAILABLE", "FEW_SEATS_AVAILABLE", "STANDING_ROOM_ONLY", "FULL"]
+    """Current occupancy level of the tram
+
+    .. versionadded:: 0.5.0
+    """
+    AVM_timestamp: datetime
+    """Timestamp of the tram vehicle driver's systems at which the tram's data in this dataset was sent
+
+    .. versionadded:: 0.5.0
+    """
 
 
 @dataclass(kw_only=True, slots=True)
@@ -146,23 +179,41 @@ class TramDestination(TramTrackerData):
 class TramStop(TramTrackerData):
     """Represents a tram stop."""
 
-    stop_id: int | None
-    """This stop's TramTracker code"""
+    stop_id: int
+    """This stop's TramTracker code
+    
+    .. versionchanged:: 0.5.0
+        Code modified so that this will no longer return ``None``
+    """
     stop_name: str
     """Name of this stop"""
-    stop_number: str | None
-    """Stop number of this stop as printed on the signage"""
-    stop_name_and_number: str | None
-    """Stop name and number combined in one string"""
+    stop_number: str
+    """Stop number of this stop as printed on the signage
+    
+    .. versionchanged:: 0.5.0
+        Code modified so that this will no longer return ``None``
+    """
     locality: str | None
-    """Locality (suburb/town) this stop is in"""
-    location: tuple[float, float] | None
-    """Currently unused; latitude-longitude coordinates of this stop"""
-    route_id: None
-    """Currently unused"""
+    """Locality (suburb/town) this stop is in; ``None`` if not provided"""
+    location: tuple[float, float]
+    """Currently unused; will always return ``(0.0, 0.0)``
+    
+    .. versionchanged:: 0.5.0
+        Value returned as-is from server (instead of substituting ``None`` for ``(0.0, 0.0)``) for consistency with :mod:`ptv_timetable` definitions
+    """
+    route_id: Literal[0]
+    """Currently unused
+    
+    .. versionchanged:: 0.5.0
+        Value returned as-is from server (instead of substituting ``None`` for ``0``) for consistency with :mod:`ptv_timetable` definitions
+    """
     destination: None
     """Currently unused"""
-    distance_to_location: float | None
-    """Currently unused"""
+    distance_to_location: float
+    """Currently unused; will always return ``0.0``
+    
+    .. versionchanged:: 0.5.0
+        Value returned as-is from server (instead of substituting ``None`` for ``0.0``) for consistency with :mod:`ptv_timetable` definitions
+    """
     city_direction: str | None
-    """Descriptor of the direction of travel for this stop (e.g. towards or away from city)"""
+    """Descriptor of the direction of travel for this stop (e.g. towards or away from city); ``None`` if not provided"""